Create TokenExpiryManager to handle refresh and remaining time to token expiry
